+++ docs/changelog.md
@@ -8,7 +8,7 @@ package will also work with version 0.6.1, but may break with version 0.7.0.
 
 ## Unreleased
 
-No changes yet!
+- Add {meth}`~cernml.mpl_utils.FigureRenderer.close()` to {class}`~cernml.mpl_utils.FigureRenderer`.
 
 ## v0.2.3